+++ libs/pandautils/schemas/com.instructure.pandautils.room.offline.OfflineDatabase/1.json
@@ -2,7 +2,7 @@
   "formatVersion": 1,
   "database": {
     "version": 1,
-    "identityHash": "1a187ca4971c65fe9154b9d46efa32b0",
+    "identityHash": "3f0779261fbeeac5cf1873b3a352c155",
     "entities": [
       {
         "tableName": "AssignmentDueDateEntity",
@@ -4292,13 +4292,19 @@
       },
       {
         "tableName": "FileSyncSettingsEntity",
-        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`id` INTEGER NOT NULL, `courseId` INTEGER NOT NULL, `url` TEXT, PRIMARY KEY(`id`), FOREIGN KEY(`courseId`) REFERENCES `CourseSyncSettingsEntity`(`courseId`) ON UPDATE NO ACTION ON DELETE CASCADE )",
+        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`id` INTEGER NOT NULL, `fileName` TEXT, `courseId` INTEGER NOT NULL, `url` TEXT, PRIMARY KEY(`id`), FOREIGN KEY(`courseId`) REFERENCES `CourseSyncSettingsEntity`(`courseId`) ON UPDATE NO ACTION ON DELETE CASCADE )",
         "fields": [
           {
             "fieldPath": "id",
             "columnName": "id",
             "affinity": "INTEGER",
             "notNull": true
+          },
+          {
+            "fieldPath": "fileName",
+            "columnName": "fileName",
+            "affinity": "TEXT",
+            "notNull": false
           },
           {
             "fieldPath": "courseId",
@@ -4814,7 +4820,7 @@
     "views": [],
     "setupQueries": [
       "CREATE TABLE IF NOT EXISTS room_master_table (id INTEGER PRIMARY KEY,identity_hash TEXT)",
-      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, '1a187ca4971c65fe9154b9d46efa32b0')"
+      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, '3f0779261fbeeac5cf1873b3a352c155')"
     ]
   }
 }
