+++ libs/pandautils/src/main/java/com/instructure/pandautils/features/offline/offlinecontent/OfflineContentRepository.kt
@@ -34,9 +34,9 @@ import com.instructure.pandautils.room.offline.model.CourseSyncSettingsWithFiles
 
 class OfflineContentRepository(
     private val coursesApi: CourseAPI.CoursesInterface,
-    private val fileFolderApi: FileFolderAPI.FilesFoldersInterface,
     private val courseSyncSettingsDao: CourseSyncSettingsDao,
-    private val fileSyncSettingsDao: FileSyncSettingsDao
+    private val fileSyncSettingsDao: FileSyncSettingsDao,
+    private val courseFileRepository: CourseFileRepository
 ) {
     suspend fun getCourse(courseId: Long): Course {
         val params = RestParams(isForceReadFromNetwork = true)
@@ -52,51 +52,9 @@ class OfflineContentRepository(
         return coursesResult.dataOrThrow.filter { it.isValidTerm() && it.hasActiveEnrollment() }
     }
 
+
     suspend fun getCourseFiles(courseId: Long): List<FileFolder> {
-        val params = RestParams(isForceReadFromNetwork = true)
-        val rootFolderResult =
-            fileFolderApi.getRootFolderForContext(courseId, CanvasContext.Type.COURSE.apiString, params)
-
-        if (rootFolderResult.isFail) return emptyList()
-
-        return getAllFiles(rootFolderResult.dataOrThrow)
-    }
-
-    private suspend fun getAllFiles(folder: FileFolder): List<FileFolder> {
-        val result = mutableListOf<FileFolder>()
-        val subFolders = getFolders(folder)
-
-        val currentFolderFiles = getFiles(folder)
-        result.addAll(currentFolderFiles)
-
-        for (subFolder in subFolders) {
-            val subFolderFiles = getAllFiles(subFolder)
-            result.addAll(subFolderFiles)
-        }
-
-        return result
-    }
-
-    private suspend fun getFolders(folder: FileFolder): List<FileFolder> {
-        val params = RestParams(usePerPageQueryParam = true, isForceReadFromNetwork = true)
-        val foldersResult = fileFolderApi.getFirstPageFolders(folder.id, params).depaginate { nextUrl ->
-            fileFolderApi.getNextPageFileFoldersList(nextUrl, params)
-        }
-
-        return foldersResult.dataOrNull.orEmpty().filterValidFileFolders()
-    }
-
-    private suspend fun getFiles(folder: FileFolder): List<FileFolder> {
-        val params = RestParams(usePerPageQueryParam = true, isForceReadFromNetwork = true)
-        val filesResult = fileFolderApi.getFirstPageFiles(folder.id, params).depaginate { nextUrl ->
-            fileFolderApi.getNextPageFileFoldersList(nextUrl, params)
-        }
-
-        return filesResult.dataOrNull.orEmpty().filterValidFileFolders()
-    }
-
-    private fun List<FileFolder>.filterValidFileFolders() = this.filter {
-        !it.isHidden && !it.isLocked && !it.isHiddenForUser && !it.isLockedForUser
+        return courseFileRepository.getCourseFiles(courseId)
     }
 
     suspend fun findCourseSyncSettings(courseId: Long): CourseSyncSettingsWithFiles {
