+++ libs/pandautils/src/test/java/com/instructure/pandautils/features/offline/offlinecontent/OfflineContentRepositoryTest.kt
@@ -45,12 +45,12 @@ import java.time.OffsetDateTime
 class OfflineContentRepositoryTest {
 
     private val coursesApi: CourseAPI.CoursesInterface = mockk(relaxed = true)
-    private val fileFolderApi: FileFolderAPI.FilesFoldersInterface = mockk(relaxed = true)
     private val courseSyncSettingsDao: CourseSyncSettingsDao = mockk(relaxed = true)
     private val fileSyncSettingsDao: FileSyncSettingsDao = mockk(relaxed = true)
+    private val courseFileRepository: CourseFileRepository = mockk(relaxed = true)
 
     private val repository =
-        OfflineContentRepository(coursesApi, fileFolderApi, courseSyncSettingsDao, fileSyncSettingsDao)
+        OfflineContentRepository(coursesApi, courseSyncSettingsDao, fileSyncSettingsDao, courseFileRepository)
 
     @Test
     fun `Returns course`() = runTest {
@@ -128,26 +128,6 @@ class OfflineContentRepositoryTest {
     }
 
     @Test
-    fun `Returns course files`() = runTest {
-        val root = FileFolder(id = 1)
-        val files = listOf(FileFolder(id = 2), FileFolder(id = 3))
-        val folders = listOf(FileFolder(id = 4), FileFolder(id = 5))
-        val subfolderFiles = listOf(FileFolder(id = 6), FileFolder(id = 7))
-
-        coEvery { fileFolderApi.getRootFolderForContext(any(), any(), any()) } returns DataResult.Success(root)
-        coEvery { fileFolderApi.getFirstPageFiles(1, any()) } returns DataResult.Success(files)
-        coEvery { fileFolderApi.getFirstPageFolders(1, any()) } returns DataResult.Success(folders)
-        coEvery { fileFolderApi.getFirstPageFiles(4, any()) } returns DataResult.Success(subfolderFiles)
-        coEvery { fileFolderApi.getFirstPageFolders(4, any()) } returns DataResult.Success(emptyList())
-        coEvery { fileFolderApi.getFirstPageFiles(5, any()) } returns DataResult.Success(emptyList())
-        coEvery { fileFolderApi.getFirstPageFolders(5, any()) } returns DataResult.Success(emptyList())
-
-        val result = repository.getCourseFiles(1)
-
-        Assert.assertEquals(files + subfolderFiles, result)
-    }
-
-    @Test
     fun `Create default course settings`() = runTest {
         val expected = CourseSyncSettingsEntity(1L, false)
 
@@ -162,7 +142,7 @@ class OfflineContentRepositoryTest {
 
     @Test
     fun `Add file calls dao`() = runTest {
-        val fileSettings = FileSyncSettingsEntity(1L, 1L, null)
+        val fileSettings = FileSyncSettingsEntity(1L, "testFile.pdf", 1L, null)
 
         repository.saveFileSettings(fileSettings)
 
@@ -190,44 +170,5 @@ class OfflineContentRepositoryTest {
         repository.updateCourseSyncSettings(1L, courseSyncSettings, emptyList())
 
         coVerify(exactly = 1) { courseSyncSettingsDao.update(courseSyncSettings) }
-    }
-
-    @Test
-    fun `Hidden and locked files and folders are filtered`() = runTest {
-        val root = FileFolder(id = 1)
-        val files = listOf(
-            FileFolder(id = 2),
-            FileFolder(id = 3, isHidden = true),
-            FileFolder(id = 4, isLocked = true),
-            FileFolder(id = 5, isHiddenForUser = true),
-            FileFolder(id = 6, isLockedForUser = true)
-        )
-        val folders = listOf(
-            FileFolder(id = 7),
-            FileFolder(id = 8, isHidden = true),
-            FileFolder(id = 9, isLocked = true),
-            FileFolder(id = 10, isHiddenForUser = true),
-            FileFolder(id = 11, isLockedForUser = true)
-        )
-        val subFolderFiles = listOf(FileFolder(id = 12))
-
-        coEvery { fileFolderApi.getRootFolderForContext(any(), any(), any()) } returns DataResult.Success(root)
-        coEvery { fileFolderApi.getFirstPageFiles(1, any()) } returns DataResult.Success(files)
-        coEvery { fileFolderApi.getFirstPageFolders(1, any()) } returns DataResult.Success(folders)
-        coEvery { fileFolderApi.getFirstPageFiles(7, any()) } returns DataResult.Success(subFolderFiles)
-        coEvery { fileFolderApi.getFirstPageFolders(7, any()) } returns DataResult.Success(emptyList())
-
-        val result = repository.getCourseFiles(1)
-
-        Assert.assertEquals(files.subList(0, 1) + subFolderFiles, result)
-    }
-
-    @Test
-    fun `Returns empty list when files request fails`() = runTest {
-        coEvery { fileFolderApi.getRootFolderForContext(any(), any(), any()) } returns DataResult.Fail()
-
-        val result = repository.getCourseFiles(1)
-
-        Assert.assertEquals(emptyList<FileFolder>(), result)
     }
 }
